Use Hotwire for Slip.js initialization
